python/webapp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.1
openai>=1.40.0

